+++ src/main/java/com/scanai/api/domain/analisepedecuba/dto/DadosCadastroAnalisePeDeCuba.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalDate;
 
-public record DadosCadastroAnalisePeDeCuba(@NotNull Long fkpedecuba, @NotNull Long fkfuncionario, @NotNull float densidade, @NotNull int temperatura) {
+public record DadosCadastroAnalisePeDeCuba(@NotNull Long fkpedecuba, @NotNull Long fkfuncionario, @NotNull Float densidade, @NotNull Float temperatura) {
 }
