+++ src/main/java/com/scanai/api/domain/analisepedecuba/dto/DadosDetalhamentoAnalisePeDeCuba.java
@@ -5,7 +5,7 @@ import com.scanai.api.domain.analisepedecuba.Analisepedecuba;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-public record DadosDetalhamentoAnalisePeDeCuba(Long id, Long fkpedecuba, Long fkfuncionario, float densidade, LocalDateTime data, int temperatura) {
+public record DadosDetalhamentoAnalisePeDeCuba(Long id, Long fkpedecuba, Long fkfuncionario, Float densidade, LocalDateTime data, Float temperatura) {
     public DadosDetalhamentoAnalisePeDeCuba(Analisepedecuba analisepedecuba) {
         this(
                 analisepedecuba.getId(),
