+++ src/main/java/com/scanai/api/domain/mostro/dto/DadosCadastroMostro.java
@@ -2,5 +2,5 @@ package com.scanai.api.domain.mostro.dto;
 
 import jakarta.validation.constraints.NotNull;
 
-public record DadosCadastroMostro(@NotNull Long fkfuncionario, @NotNull int volume, Long fkmostro1, Long fkmostro2) {
+public record DadosCadastroMostro(@NotNull Long fkfuncionario, @NotNull Float volume, Long fkmostro1, Long fkmostro2) {
 }
